+++ R/rct.R
@@ -12,7 +12,8 @@
 #' @examples
 #' rct(1, 2, 3, 4)
 #'
-rct <- function(xmin = double(), ymin = double(), xmax = double(), ymax = double(), crs = wk_crs_auto()) {
+rct <- function(xmin = double(), ymin = double(), xmax = double(), ymax = double(),
+                crs = wk_crs_auto(), geodesic = FALSE) {
   vec <- new_wk_rct(
     recycle_common(
       xmin = as.double(xmin),
@@ -20,7 +21,8 @@ rct <- function(xmin = double(), ymin = double(), xmax = double(), ymax = double
       xmax = as.double(xmax),
       ymax = as.double(ymax)
     ),
-    crs = wk_crs_auto_value(xmin, crs)
+    crs = wk_crs_auto_value(xmin, crs),
+    geodesic = if (isTRUE(geodesic)) TRUE else NULL
   )
 
   validate_wk_rct(vec)
@@ -72,8 +74,8 @@ validate_wk_rct <- function(x) {
 #' @export
 #'
 new_wk_rct <- function(x = list(xmin = double(), ymin = double(), xmax = double(), ymax = double()),
-                       crs = NULL) {
-  structure(x, class = c("wk_rct", "wk_rcrd"), crs = crs)
+                       crs = NULL, geodesic = NULL) {
+  structure(x, class = c("wk_rct", "wk_rcrd"), crs = crs, geodesic = geodesic)
 }
 
 #' @export
@@ -91,5 +93,9 @@ format.wk_rct <- function(x, ...) {
   replacement <- as_rct(value)
   result <- Map("[<-", unclass(x), i, unclass(replacement))
   names(result) <- c("xmin", "ymin", "xmax", "ymax")
-  new_wk_rct(result, crs = wk_crs_output(x, replacement))
+  new_wk_rct(
+    result,
+    crs = wk_crs_output(x, replacement),
+    geodesic = if (wk_is_geodesic_output(x, replacement)) TRUE else NULL
+  )
 }